Cut main-thread tile apply cost: split-by-class moves to the worker, drop the buildingIndex copy, time-budget the drain loop
Three changes that together address the per-tile main-thread stall during
initial load on big viewports (16 fps on a Mac with eGPU was traceable to
the apply phase, not the GPU):

1. splitByClass runs in the worker. The landuse/roads/rails extractors
   each emit a per-vertex class attribute; the main thread used to walk
   the index buffer and rebuild remapped sub-geometries per tile, per
   layer. That work now happens in the worker post-extract and the
   decoded payload carries pre-split `submeshes` typed arrays straight
   to a BufferGeometry on the main thread. New: SubmeshGeometry type +
   splitByClass helper under tiles/worker/shared/.

2. buildingIndex ships as Float32Array. The worker used to emit Uint32
   and the BuildingsLayer copied it to Float32 (shader expects float).
   Worker now emits Float32 directly — zero copy on the main thread.

3. Apply pacing is a 6ms time budget per frame instead of a fixed
   "N tiles per frame" count. The old count budget worked fine for
   sparse tiles but a fat tile (7 layers + extruded buildings) blew the
   frame budget several times over. The new loop applies tiles
   closest-to-camera-target first and yields once the budget is spent,
   while always applying at least one tile per frame so the queue can
   drain when a single build exceeds the budget.

Public API: `performance.maxTileBuildsPerFrame` is renamed to
`performance.maxTileApplyMsPerFrame`. Docs updated. No other behavior
change.

+++ src/tiles/worker/extractors/buildings.ts
@@ -319,7 +319,10 @@ export function extractBuildings(
     positions: new Float32Array(positions),
     indices: new Uint32Array(indices),
     normals: new Float32Array(normals),
-    attributes: { buildingIndex: new Uint32Array(buildingIndex) },
+    // Float32 matches the shader's `attribute float buildingIndex` exactly,
+    // so the main thread can wrap this in a BufferAttribute with zero copy.
+    // Building indices fit well within Float32's 24-bit mantissa precision.
+    attributes: { buildingIndex: new Float32Array(buildingIndex) },
     metadata: { totalVolume, buildings }
   };
 }
